Move BaseEndpointAuth
