+++ src/frontend/packages/cloud-foundry/src/cf-entity-generator.ts
@@ -4,9 +4,9 @@ import { combineLatest, Observable, of } from 'rxjs';
 import { first, map } from 'rxjs/operators';
 
 import { urlValidationExpression } from '../../core/src/core/utils.service';
-import { BaseEndpointAuth } from '../../core/src/features/endpoints/endpoint-auth';
 import { AppState, GeneralEntityAppState } from '../../store/src/app-state';
 import { metricEntityType } from '../../store/src/base-entity-schemas';
+import { BaseEndpointAuth } from '../../store/src/endpoint-auth';
 import {
   StratosBaseCatalogEntity,
   StratosCatalogEndpointEntity,
