+++ src/frontend/packages/core/src/features/endpoints/connect-endpoint/connect-endpoint.component.ts
@@ -14,11 +14,11 @@ import {
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Subscription } from 'rxjs';
 
+import { BaseEndpointAuth } from '../../../../../store/src/endpoint-auth';
 import { entityCatalog } from '../../../../../store/src/entity-catalog/entity-catalog';
 import { EndpointAuthTypeConfig, IAuthForm, IEndpointAuthComponent } from '../../../../../store/src/extension-types';
 import { safeUnsubscribe } from '../../../core/utils.service';
 import { ConnectEndpointConfig, ConnectEndpointData, ConnectEndpointService } from '../connect.service';
-import { BaseEndpointAuth } from '../endpoint-auth';
 
 @Component({
   selector: 'app-connect-endpoint',
