+++ src/frontend/packages/store/src/endpoint-auth.ts
@@ -1,10 +1,16 @@
 import { Validators } from '@angular/forms';
 
-import { EndpointAuthTypeConfig, EndpointType } from '../../../../store/src/extension-types';
-import { CredentialsAuthFormComponent } from './connect-endpoint-dialog/auth-forms/credentials-auth-form.component';
-import { NoneAuthFormComponent } from './connect-endpoint-dialog/auth-forms/none-auth-form.component';
-import { SSOAuthFormComponent } from './connect-endpoint-dialog/auth-forms/sso-auth-form.component';
-import { EndpointAuthTypeNames } from './endpoint-helpers';
+import {
+  CredentialsAuthFormComponent,
+} from '../../core/src/features/endpoints/connect-endpoint-dialog/auth-forms/credentials-auth-form.component';
+import {
+  NoneAuthFormComponent,
+} from '../../core/src/features/endpoints/connect-endpoint-dialog/auth-forms/none-auth-form.component';
+import {
+  SSOAuthFormComponent,
+} from '../../core/src/features/endpoints/connect-endpoint-dialog/auth-forms/sso-auth-form.component';
+import { EndpointAuthTypeNames } from '../../core/src/features/endpoints/endpoint-helpers';
+import { EndpointAuthTypeConfig, EndpointType } from './extension-types';
 
 export abstract class BaseEndpointAuth {
   static readonly UsernamePassword = {
